requirements.txt
seaborn==0.9.0
pandas==0.24.2
numpy==1.16.4
sklearn_pandas==1.8.0
geopy==1.20.0
plotly==4.4.1
dython==0.3.1
scipy==1.1.0
statsmodels==0.9.0
ipython==7.11.0
matplotlib==3.1.2
scikit_learn==0.22
nbformat==4.2.0
